pyproject.toml
[project]
name = "filmgraph"
version = "0.1.0"
description = "A minimal, structured data format for film analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Chapter", email = "hello@chapter.film" },
]
keywords = ["film", "cinema", "video", "analysis", "pydantic", "schema"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
    "Typing :: Typed",
]
dependencies = ["pydantic>=2.0"]

[project.optional-dependencies]
docx = ["python-docx>=1.1"]
otio = ["opentimelineio>=0.17"]
all = ["python-docx>=1.1", "opentimelineio>=0.17"]

[project.urls]
Homepage = "https://github.com/Chapter41/filmgraph"
Documentation = "https://Chapter41.github.io/filmgraph"
Repository = "https://github.com/Chapter41/filmgraph"
Issues = "https://github.com/Chapter41/filmgraph/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/filmgraph"]
